printf v0.5

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: goliano- <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/07 13:06:05 by goliano-          #+#    #+#             */
-/*   Updated: 2021/09/07 13:50:47 by goliano-         ###   ########.fr       */
+/*   Updated: 2021/09/10 16:30:18 by goliano-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 # include <stdio.h>
 # include <stdint.h>
 
-char	*to_hex(uintptr_t n, int t);
+char	*to_hex(unsigned long int n, int t);
 void    ft_putunbr(long n);
 
 #endif
